requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-RESTful==0.3.9
PyMySQL==1.0.2
python-dotenv==0.19.0
asgiref==3.4.1
flask-cors==3.0.10